Checkpoint 22-03-2026 02:31

+++ frontend/src/services/client/tool-services.ts
@@ -1,5 +1,6 @@
 import { ToolSchema } from "@/domain/schemas/ToolSchema";
 import { apiClient } from "./api-client";
+import { ToolFilters } from "@/domain/schemas/ToolFilters";
 
 export const addToolService = async(tool: ToolSchema) =>{
     const api = await apiClient();
@@ -13,6 +14,21 @@ export const updateToolService = async(id:string, tool: ToolSchema) =>{
 }
 export const deleteToolByRouteService = async(id:string) =>{
     const api = await apiClient();
-    const response = await api.delete(`/api/tools/${id}`);
+    const response = await api.delete(`/api/admin/tools/${id}`);
+    return response;
+}
+export const getToolsByPagination = async(filters: ToolFilters) =>{
+    const api = await apiClient();
+    const params = new URLSearchParams();
+    params.append("page", filters.page.toString());
+    if(filters.search){
+        params.append("search", filters.search);
+    }
+    const response = await api.get(`/api/admin/tools/pagination?${params}`);
+    return response;
+}
+export const getToolByIdService = async(id:string) =>{
+    const api = await apiClient();
+    const response = await api.get(`/api/admin/tools/${id}`);
     return response;
 }
